fix: Issue #22
Change the name of the contextSegments array to optionsSyntax and update properties to
allow it to be used for validation of all options properties provided by
gitaclue.get(options) callers.

Note that this is a "safety" commit prior to making more extensive changes needed to
resolve this issue.

Resolves: https://github.com/jdmedlock/GitAClue/issues/22
See also: N/a